Refactor owners parser
